refactor: Update PeerContextProvider to include createAnswer function
The PeerContextProvider component has been updated to include the createAnswer function. This function allows for the creation of an answer in response to an offer received from another peer. It sets the remote description, creates the answer, and sets the local description accordingly. This addition enhances the functionality of the PeerContextProvider for establishing WebRTC connections.