refactor: Update Lobby component to include video streaming functionality
The Lobby component has been refactored to include video streaming functionality using the getUserMedia API. This allows users to access their camera and microphone and stream their video and audio in the lobby. The stream is then displayed using the ReactPlayer component. This enhancement improves the functionality of the Lobby component for establishing WebRTC connections and enhances the user experience.