refactor: Update PeerContextProvider to include sendStream function
The PeerContextProvider component has been updated to include the sendStream function. This function allows for sending the local stream to the remote peer by adding each track of the stream to the RTCPeerConnection. This enhancement improves the functionality of the PeerContextProvider for establishing WebRTC connections and enables video streaming between peers.