refactor: Update Lobby component to include remote email ID
The Lobby component has been updated to include the display of the remote email ID in the UI. This allows users to see the email ID of the peer they are connected to in the lobby. This enhancement improves the user experience and provides better context during video streaming sessions.